+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Update images implementation [#4](https://github.com/fdrandolfi/composition-tools-frontend/pull/4)
 - Add commons docs, favicons & Update Scaletor [#3](https://github.com/fdrandolfi/composition-tools-frontend/pull/3)
 
 ## [1.0.0]